spo ambient_light

+++ include/objects.h
@@ -6,13 +6,13 @@
 /*   By: denizozd <denizozd@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/03 15:57:56 by denizozd          #+#    #+#             */
-/*   Updated: 2024/05/03 16:06:58 by denizozd         ###   ########.fr       */
+/*   Updated: 2024/05/06 14:59:33 by denizozd         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 typedef struct s_amblight{
 	char	*name;
-	float	intensity;
+	double	intensity;
 	int		r;
 	int		g;
 	int		b;
